+++ docs/examples/data_transfer_objects/dto_auto_conversion.py
@@ -4,7 +4,7 @@ from sqlalchemy import Column, Float, Integer, String
 from sqlalchemy.orm import Mapped, declarative_base
 
 from starlite import DTOFactory, HTTPException, Starlite, get
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
+from starlite.plugins.sqlalchemy import SQLAlchemyPlugin
 from starlite.status_codes import HTTP_404_NOT_FOUND
 
 sqlalchemy_plugin = SQLAlchemyPlugin()
