+++ docs/examples/datastructures/pagination/using_offset_pagination_with_sqlalchemy.py
@@ -5,7 +5,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import Mapped, declarative_base
 
 from starlite import AbstractAsyncOffsetPaginator, OffsetPagination, Provide, Starlite, get
-from starlite.plugins.sql_alchemy import SQLAlchemyConfig, SQLAlchemyPlugin
+from starlite.plugins.sqlalchemy import SQLAlchemyConfig, SQLAlchemyPlugin
 
 Base = declarative_base()
