+++ docs/examples/plugins/sqlalchemy_plugin/sqlalchemy_relationships.py
@@ -4,7 +4,7 @@ from sqlalchemy import Column, Float, ForeignKey, Integer, String, create_engine
 from sqlalchemy.orm import Mapped, Session, declarative_base, relationship
 
 from starlite import HTTPException, Starlite, get
-from starlite.plugins.sql_alchemy import SQLAlchemyConfig, SQLAlchemyPlugin
+from starlite.plugins.sqlalchemy import SQLAlchemyConfig, SQLAlchemyPlugin
 from starlite.status_codes import HTTP_404_NOT_FOUND
 
 engine = create_engine("sqlite+pysqlite://")
