+++ docs/examples/plugins/sqlalchemy_plugin/sqlalchemy_relationships_to_many.py
@@ -5,7 +5,7 @@ from sqlalchemy import Column, ForeignKey, Integer, String, create_engine, selec
 from sqlalchemy.orm import Mapped, Session, declarative_base, relationship
 
 from starlite import HTTPException, Starlite, get
-from starlite.plugins.sql_alchemy import SQLAlchemyConfig, SQLAlchemyPlugin
+from starlite.plugins.sqlalchemy import SQLAlchemyConfig, SQLAlchemyPlugin
 from starlite.status_codes import HTTP_404_NOT_FOUND
 
 engine = create_engine("sqlite+pysqlite:///test.sqlite")
