+++ docs/examples/plugins/sqlalchemy_plugin/sqlalchemy_sync.py
@@ -4,7 +4,7 @@ from sqlalchemy import Column, Float, Integer, String, select
 from sqlalchemy.orm import Mapped, Session, declarative_base
 
 from starlite import DTOFactory, HTTPException, Starlite, get, post
-from starlite.plugins.sql_alchemy import SQLAlchemyConfig, SQLAlchemyPlugin
+from starlite.plugins.sqlalchemy import SQLAlchemyConfig, SQLAlchemyPlugin
 from starlite.status_codes import HTTP_404_NOT_FOUND
 
 Base = declarative_base()
