--- docs/reference/plugins/sql_alchemy.rst
@@ -1,5 +1,0 @@
-sql_alchemy
-===========
-
-.. automodule:: starlite.plugins.sql_alchemy
-    :members:
