+++ docs/reference/plugins/sqlalchemy.rst
@@ -1,0 +1,5 @@
+sqlalchemy
+===========
+
+.. automodule:: starlite.plugins.sqlalchemy
+    :members:
