+++ mypy.ini
@@ -12,10 +12,10 @@ disallow_any_generics = False
 implicit_reexport = False
 show_error_codes = True
 
-[mypy-starlite.plugins.sql_alchemy.plugin]
+[mypy-starlite.plugins.sqlalchemy.plugin]
 disable_error_code = dict-item
 
-[mypy-tests.plugins.sql_alchemy_plugin.test_sql_alchemy_table_types]
+[mypy-tests.plugins.sqlalchemy_plugin.test_sqlalchemy_table_types]
 disable_error_code = arg-type
 
 [mypy-tests.*]
