+++ pyproject.toml
@@ -94,7 +94,7 @@ python-dotenv = "*"
 python-jose = "*"
 redis = "*"
 rich = "*"
-sqlalchemy = "<2.0.0"
+sqlalchemy = "*"
 starlette = "*"
 structlog = "*"
 tortoise-orm = ">=0.17.0"
