+++ tests/dto_factory/test_dto_factory_partials.py
@@ -1,5 +1,5 @@
 from starlite import DTOFactory
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
+from starlite.plugins.sqlalchemy import SQLAlchemyPlugin
 from starlite.types.partial import Partial
 from tests import Car, Person
