+++ tests/plugins/sqlalchemy_plugin/test_sqlalchemy_plugin_integration.py
@@ -6,10 +6,10 @@ from pydantic_factories.value_generators.primitives import (
 )
 
 from starlite import get, post
-from starlite.plugins.sql_alchemy import SQLAlchemyPlugin
+from starlite.plugins.sqlalchemy import SQLAlchemyPlugin
 from starlite.status_codes import HTTP_200_OK, HTTP_201_CREATED
 from starlite.testing import create_test_client
-from tests.plugins.sql_alchemy_plugin import Company
+from tests.plugins.sqlalchemy_plugin import Company
 
 companies = [
     Company(id=i, name=create_random_string(min_length=5, max_length=20), worth=create_random_float(minimum=1))
@@ -39,7 +39,7 @@ def bulk_create_company(data: List[Company]) -> List[Company]:
     return data
 
 
-def test_return_single_sql_alchemy_model_instances() -> None:
+def test_return_single_sqlalchemy_model_instances() -> None:
     with create_test_client([get_companies], plugins=[SQLAlchemyPlugin()]) as client:
         response = client.get("/companies")
         assert response.status_code == HTTP_200_OK
@@ -48,7 +48,7 @@ def test_return_single_sql_alchemy_model_instances() -> None:
         ]
 
 
-def test_return_of_a_single_sql_alchemy_model_instance() -> None:
+def test_return_of_a_single_sqlalchemy_model_instance() -> None:
     with create_test_client([get_company_by_id], plugins=[SQLAlchemyPlugin()]) as client:
         response = client.get("/companies/1")
         assert response.status_code == HTTP_200_OK
@@ -61,7 +61,7 @@ def test_return_of_a_single_sql_alchemy_model_instance() -> None:
         )
 
 
-def test_serializing_a_single_sql_alchemy_instance() -> None:
+def test_serializing_a_single_sqlalchemy_instance() -> None:
     company = {"id": 10, "name": create_random_string(), "worth": create_random_float()}
     with create_test_client([create_company], plugins=[SQLAlchemyPlugin()]) as client:
         response = client.post("/companies", json=company)
@@ -69,7 +69,7 @@ def test_serializing_a_single_sql_alchemy_instance() -> None:
         assert response.json() == company
 
 
-def test_serializing_a_list_of_sql_alchemy_instances() -> None:
+def test_serializing_a_list_of_sqlalchemy_instances() -> None:
     serialized_companies = [{"id": company.id, "name": company.name, "worth": company.worth} for company in companies]
     with create_test_client([bulk_create_company], plugins=[SQLAlchemyPlugin()]) as client:
         response = client.post("/companies/bulk", json=serialized_companies)
